phishing-url-detector/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-jwt-extended==4.5.3

# Machine Learning (Python 3.13 compatible)
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# URL Processing & Feature Extraction
tldextract==5.1.1
python-whois==0.8.0
dnspython==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
validators==0.22.0

# SSL/Certificate Analysis
cryptography==41.0.7
pyOpenSSL==23.3.0

# Threat Intelligence APIs
virustotal-python==1.0.2
shodan==1.31.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Alerting
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
colorama==0.4.6
tqdm==4.66.1
